[#141]: Make sure that Objects that are not part of currently selected render layer are not being selectable

+++ editor/gui/editor_gui.cpp
@@ -918,7 +918,7 @@ EditorGUI::RenderGameObjectMenu() // NOLINT
                   if (ImGui::Selectable(item.c_str()))
                   {
                      parent_.AddToWorkQueue([this, item] {
-                        const auto layer = static_cast< uint32_t >(std::stoi(item));
+                        const auto layer = std::stoi(item);
                         const auto oldLayer =
                            currentlySelectedGameObject_->GetSprite().GetRenderInfo().layer;
                         currentlySelectedGameObject_->GetSprite().ChangeRenderLayer(layer);
